Sort works!  With sort spec.
I also started doing some benchmarking against Kettle.  In general,
read/writes are about the same (datasets might even be a bit faster
than serialized kettle files!)  If the sort is done in-memory, it's only about
50% slower than Kettle.  Kettle is multi-threaded, so I think this is a very
favorable comparison.  However, when I need to split the file and do an
external sort, it's almost 3 times slower than kettle.  This seems to
be an issue the interleave method, which needs some work.