Ugh..... I'm going to revisit the delegator methods I'm using in Variable and VariableSet.
The problem is that the way I have it set up now, I can't use methods defined outside of
the scope of VariableSet.modify! block.  This is not ok.  I'm pretty sure there's a way around
it, but at this point, I think I'm just going to make the method calls more explicit, like I had before.
It's a little more verbose, but it's safer.  Maybe I can have this cleaner DSL syntax in Remi 2.0.