Adds a Now formula for times
Also, I monkey-patched Time.current so that the value gets memoized the first
time it is called IN A TEST.  This is a workaround to the issue that when Time.current gets
called in the job may be slightly different when Time.current gets called in the test.
Hope this isn't a bad idea.