S3 Extractor support
This commit adds support for an S3 extractor type.  I've
taking the original sftp extractor and abstracted most of
the methods into an FileSystem type.  However, I did not
replace the existing class because the abstraction required
some breaking changes and I need time to resolve them in
various jobs.