Resolve a bug with parsing google sheet data
When the last cell of a row was blank, all populated rows below that
in the same column were being shifted up.  This is a consequence of
the google sheet API returning a row as an array that has a size up
to the last populated column in a row.