FIX - Subtransforms map to the correct source index
This bug affected cases where the source and target of
a subtransform referred to the same subject in the parent job.
In these cases, if the index was modified (e.g., by sorting or joining), the
values that would get put back onto the parent job datasubject could be
out of order.