DE-1090 - Allow users to define values that don't get type enforced
We needed this to support cases where empty data type besides nil.  For example,
suppose we have a data target that is mostly expecting dates in a particular field.
If we don't want the field to update, we can't use nil, because that might indicate that
the field should be populated with nil.  Instead, we can define a custom class that
the data target can detect and send a special message to the target to indicate that
there is to be no update of the existing value.